requirements.txt
# Web Crawling
requests==2.31.0
beautifulsoup4==4.12.2
lxml==5.1.0
urllib3==2.1.0

# Web Framework
fastapi==0.109.0
uvicorn==0.27.0
jinja2==3.1.3

# Data Processing
pandas==2.1.4
numpy==1.26.3

# Database
sqlalchemy==2.0.25

# Text Processing
nltk==3.8.1

# Testing
pytest==7.4.4
pytest-cov==4.1.0

# Development
black==24.1.1
flake8==7.0.0
python-dotenv==1.0.0
